* Fix update-launchpad-translations to cope with no-longer-existant   slideshows. * Update translations from Launchpad. * Remove translations not present in localechooser, and thus not   available during install.

+++ update-launchpad-translations.sh
@@ -36,6 +36,7 @@ tar -C $WORKING_DIR -zxvf "$1" 1>/dev/null
 save="$pwd"
 cd $WORKING_DIR/po
 for distro in *; do
+	[ -d "$repo/po/$distro" ] || continue
 	cd "$distro"
 	rename 's/.*-//' *.po
 	rm -f *.pot
